+++ examples/webpack/package.json
@@ -1,12 +1,12 @@
 {
   "name": "webpack-pdf.js-example",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "scripts": {
     "build": "webpack"
   },
   "devDependencies": {
-    "webpack": "^5.11.1",
-    "webpack-cli": "^4.3.1",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
     "pdfjs-dist": "../../node_modules/pdfjs-dist"
   }
 }
