+++ examples/webpack/webpack.config.js
@@ -6,8 +6,8 @@ const path = require("path");
 module.exports = {
   context: __dirname,
   entry: {
-    main: "./main.js",
-    "pdf.worker": "pdfjs-dist/build/pdf.worker.entry",
+    main: "./main.mjs",
+    "pdf.worker": "pdfjs-dist/build/pdf.worker.mjs",
   },
   mode: "none",
   output: {
